modulo 6 exptions finalizado

+++ src/alunos/alunos.service.ts
@@ -1,9 +1,10 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { CreateAlunoDto } from './dto/create-aluno.dto';
 import { UpdateAlunoDto } from './dto/update-aluno.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Aluno } from './entities/aluno.entity';
 import { Repository } from 'typeorm';
+import { AlunoNotFoundException } from './exceptions/AlunoNotFoundException';
 
 @Injectable()
 export class AlunosService {
@@ -33,14 +34,32 @@ export class AlunosService {
   }
 
   async findOne(id: number) {
-    return await this.repository.findOneBy({ id: id });
+    try{
+      return await this.repository.findOneByOrFail({id: id});
+    }catch(error){
+      console.log(error.name);
+      if(error.name === 'EntityNotFoundError'){
+        throw new HttpException(
+          'Aluno não encontrado',HttpStatus.NOT_FOUND
+        );
+      }
+      throw new HttpException(
+        'Erro Interno servidor Offline', HttpStatus.INTERNAL_SERVER_ERROR
+      );
+    }
   }
 
   async update(id: number, updateAlunoDto: UpdateAlunoDto) {
-    return await this.repository.update(id, updateAlunoDto);
+      return await this.repository.update(id, updateAlunoDto);
   }
 
   async remove(id: number) {
+    const response = await this.repository.delete({id});
+
+    if( response.affected === 0){
+      throw new AlunoNotFoundException();
+    }
+
     return await this.repository.delete({id});
   }
 }
